requirements.txt
# Update your requirements.txt file
flask>=2.0.1,<3.0.0
flask-sqlalchemy>=3.0.0
psycopg2-binary>=2.9.1
werkzeug>=2.0.0,<3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
loguru==0.7.0
sqlalchemy-utils>=0.41.1

git+https://github.com/carousell/ct-grpc-python@pkg/ct-data-logic-multimodal-ad-listing/v1.3.0#subdirectory=pkg/ct-data-logic-multimodal-ad-listing